sci_tu_dashboard/requirements.txt
aioredis==1.3.1
asgiref==3.4.0
astroid==2.8.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
coverage==5.5
cryptography==3.4.7
daphne==3.0.2
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-renderer==1.9.1
dash-table==4.12.0
Django==3.2.4
django-bootstrap4==3.0.1
django-plotly-dash==1.6.4
django-redis==5.0.0
djangorestframework==3.12.4
dpd-components==0.1.0
dpd-static-support==0.0.5
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==2.1.1
incremental==21.3.0
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
mysqlclient==2.0.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
Pillow==8.3.1
platformdirs==2.4.0
plotly==5.1.0
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.11.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tenacity==8.0.1
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
zope.interface==5.4.0
